Simplify HTML/CSS code Part 1
Working on making the markup simpler so it is easier to read and modify.
This is mostly done be starting from scratch. Now I have a final product
in mind it makes it easier to keep the content semantically correct.

Betting splitting of components will be done as well.

So far just the main-layout has been really modified. The side-nav was
moved all the way to the side. It also now has a fixed size.