Restart Project from Vue Webpack template
Started the project from scratch using the provided template.
I was trying to learn the basics of webpack when I first did
the website and I feel I understand the concepts, so I wanted
to try using a proper installation of it this time.

This commit is just a basic website that has routing (But no
content)